Move Select & Select2 to flat structure
