+++ packages/client/src/components/form/DocumentUploadfield/DocumentUploaderWithOption.tsx
@@ -9,7 +9,7 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-import { ISelectOption, Select } from '@opencrvs/components/lib/forms'
+import { ISelectOption, Select } from '@opencrvs/components/lib/Select'
 import { ImageUploader } from '@opencrvs/components/lib/ImageUploader'
 import { ErrorText } from '@opencrvs/components/lib/ErrorText'
 import { DocumentPreview } from '@client/components/form/DocumentUploadfield/DocumentPreview'
