+++ packages/client/src/components/form/FormFieldGenerator.tsx
@@ -13,11 +13,11 @@ import * as React from 'react'
 import {
   CheckboxGroup,
   RadioGroup,
-  Select,
   TextArea,
   TextInput,
   RadioSize
 } from '@opencrvs/components/lib/forms'
+import { Select } from '@opencrvs/components/lib/Select'
 import { DateField } from '@opencrvs/components/lib/DateField'
 import { WarningMessage } from '@opencrvs/components/lib/WarningMessage'
 import { Link } from '@opencrvs/components/lib/typography'
