+++ packages/client/src/views/Settings/items/Language.tsx
@@ -31,7 +31,7 @@ import {
 } from '@client/views/Settings/items/components'
 import { useSelector, useDispatch } from 'react-redux'
 import { TertiaryButton, PrimaryButton } from '@opencrvs/components/lib/buttons'
-import { Select } from '@opencrvs/components/lib/forms'
+import { Select } from '@opencrvs/components/lib/Select'
 import { getAvailableLanguages } from '@client/i18n/utils'
 import { getUserDetails } from '@client/profile/profileSelectors'
 import { changeLanguage as changeLanguageActionCreator } from '@client/i18n/actions'
