+++ packages/components/src/components/forms/InputField/InputField.stories.tsx
@@ -13,7 +13,7 @@ import React from 'react'
 import { Meta, Story } from '@storybook/react'
 import { ValidIndicator, VerifyingIndicator } from '..'
 import { ArrowWithGradient } from '../../icons'
-import { Select } from '../Select'
+import { Select } from '../../Select'
 import { TextInput } from '../TextInput'
 import { InputField, IInputFieldProps } from './InputField'
 import { DateField } from '../../DateField'
